+++ src/index.spec.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
 import { render } from 'react-dom';
+import renderer from 'react-test-renderer';
 import { App } from './components/App/App';
 
 jest.mock("react-dom", () => ({ render: jest.fn() }));
@@ -15,5 +16,12 @@ describe('index', () => {
       <App />,
       document.getElementById('root')
     );
+  });
+
+  it('should render correctly', () => {
+    const tree = renderer
+      .create(<App />)
+      .toJSON();
+    expect(tree).toMatchSnapshot();
   });
 });